% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/management_functions.R
\name{micRoclean}
\alias{micRoclean}
\title{micRoclean
Primary function for decontamination low biomass microbiome data}
\usage{
micRoclean(
  counts,
  meta,
  research_goal,
  control_name,
  control_order = NA,
  blocklist,
  remove_if = 1,
  step2_threshold = 0.5,
  seed = 42
)
}
\arguments{
\item{counts}{Count matrix with samples as rows and features as counts}

\item{meta}{Data frame with columns is_control, sample_type, and (optional) sample_well}

\item{research_goal}{Define primary research goal. Options: 'biomarker' or 'orig.composition'}

\item{control_name}{Character name of controls within metadata}

\item{control_order}{Vector ordering run of sample_type controls, default NA}

\item{blocklist}{Vector of known previously identified contaminant features}

\item{remove_if}{Threshold for number of steps feature must be identified as potential contaminant to be removed from final cleaned count matrix. Default set to 1.}

\item{step2_threshold}{Threshold value for prevalence method of decontam}

\item{seed}{Random seed}
}
\value{
List object with original matrix, decontaminated matrix, character string of all removed contaminants (if 'biomarker'), and corresponding
filtering loss (FL) statistics
}
\description{
micRoclean
Primary function for decontamination low biomass microbiome data
}
