% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline2.R
\name{pipeline2}
\alias{pipeline2}
\title{pipeline2
Pipeline 2 is best used when the main goal is to identify potential biomarkers, rather than characterize the original
composition.}
\usage{
pipeline2(
  counts,
  meta,
  blocklist,
  control_name,
  remove_if = 1,
  step2_threshold = 0.5,
  seed = 42
)
}
\arguments{
\item{counts}{Count matrix with samples as rows and features as counts}

\item{meta}{dataframe with columns is_control, sample_type}

\item{blocklist}{Vector of known previously identified contaminant features}

\item{control_name}{Character name of controls within metadata}

\item{remove_if}{Threshold for number of steps feature must be identified as potential contaminant to be removed from final cleaned count matrix. Default set to 1.}

\item{step2_threshold}{Threshold value for prevalence method of decontam}

\item{seed}{Random seed}
}
\value{
List object with contaminant matrix, decontaminated count matrix, character string of all removed contaminants,
and filtering loss (FL) statistic
}
\description{
pipeline2
Pipeline 2 is best used when the main goal is to identify potential biomarkers, rather than characterize the original
composition.
}
\seealso{
Other pipeline2: 
\code{\link{step1}()},
\code{\link{step2}()},
\code{\link{step4}()}
}
\concept{pipeline2}
